requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
python-decouple==3.8
openpyxl==3.1.5
reportlab==4.2.5
psycopg[binary]
gunicorn==22.0.0
django-simple-history==3.7.0
whitenoise==6.7.0
# LDAP / Active Directory (instalar solo si se necesita autenticacion AD)
python-ldap==3.4.4
django-auth-ldap==4.8.0
# Tareas asincronas (para sincronizacion nocturna LDAP)
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1
